Major refactoring & documentation
